+++ flumine/baseflumine.py
@@ -11,7 +11,7 @@ from .worker import BackgroundWorker
 from .clients.baseclient import BaseClient
 from .markets.markets import Markets
 from .markets.market import Market
-from .markets.middleware import Middleware
+from .markets.middleware import Middleware, SimulatedMiddleware
 from .execution.betfairexecution import BetfairExecution
 from .execution.simulatedexecution import SimulatedExecution
 from .order.process import process_current_orders
@@ -73,10 +73,11 @@ class BaseFlumine:
     def run(self) -> None:
         raise NotImplementedError
 
-    def add_strategy(self, strategy: BaseStrategy) -> None:
+    def add_strategy(self, strategy: BaseStrategy, client: BaseClient = None) -> None:
         logger.info("Adding strategy {0}".format(strategy))
+        _client = client or self.client
         self.streams(strategy)  # create required streams
-        self.strategies(strategy)  # store in strategies
+        self.strategies(strategy, _client)  # store in strategies
         self.log_control(events.StrategyEvent(strategy))
 
     def add_worker(self, worker: BackgroundWorker) -> None:
@@ -295,6 +296,9 @@ class BaseFlumine:
             config.simulated = True
         else:
             config.simulated = False
+        # middleware
+        if self.BACKTEST or self.client.paper_trade:
+            self.add_market_middleware(SimulatedMiddleware())
         # login
         self.client.login()
         self.client.update_account_details()